Try to automatically reconnect to an open IMAP mailbox on error
For the Context, change cmd_handle_fatal() to flag idata->status on
the error, but not close the mailbox and disconnect.

Then have imap_check_mailbox() reconnect when IMAP_REOPEN_ALLOW is
set(so the index is prepared to rebuild with all new headers).  Try
to merge flag, envelope, and attach_del changes back in.  Replace the
existing Context pointer with the new values, and swap out the idata
passed in on success, so we don't continue to work with the wrong
idata.

The imap_check_mailbox() changes are purposely coded loosely to try
and catch all sorts of errors - not just polling and imap_idle
errors.

Create a new check_mailbox() return code to indicate a reconnect.
Display a more appropriate message in the index, but otherwise treat
it the same as a REOPEN.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/44711a243ee8fe7dc5e206d0175a5026599275ef
Co-authored-by: Richard Russon <rich@flatcap.org>